hw3

+++ models/src/main/java/edu/java/models/Request/LinkUpdate.java
@@ -2,9 +2,9 @@ package edu.java.models.Request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record LinkUpdate(@JsonProperty("id") int id,
+public record LinkUpdate(@JsonProperty("id") long id,
                          @JsonProperty("url") String url,
                          @JsonProperty("description") String description,
-                         @JsonProperty("tgChatIds") int[] tgChatIds) {
+                         @JsonProperty("tgChatIds") long[] tgChatIds) {
 
 }
