+++ models/src/main/java/edu/java/models/Response/LinkResponse.java
@@ -2,6 +2,6 @@ package edu.java.models.Response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record LinkResponse(@JsonProperty("id") int id,
+public record LinkResponse(@JsonProperty("id") long id,
                             @JsonProperty("url") String url) {
 }
